+++ BubbleBobble/BufferBdata.cpp
@@ -3,10 +3,10 @@
 
 using namespace ieg;
 
-const size_t BufferBdata::mLevelDataOffset{ 0 };
-const size_t BufferBdata::mLevelByteWidth{ 4 };
-const size_t BufferBdata::mLevelWidth{ 32 };
-const size_t BufferBdata::mLevelHeight{ 25 };
+const int BufferBdata::mLevelDataOffset{ 0 };
+const int BufferBdata::mLevelByteWidth{ 4 };
+const int BufferBdata::mLevelWidth{ 32 };
+const int BufferBdata::mLevelHeight{ 25 };
 
 BufferBdata::BufferBdata(const std::string& filename)
 	: Buffer(filename)
@@ -40,16 +40,16 @@ bool BufferBdata::LoadFile()
 	return true;
 }
 
-void BufferBdata::GetLayout(char* pLayout, size_t level)
+void BufferBdata::GetLayout(char* pLayout, int level)
 {
-	for (size_t row{ 0 }; row < mLevelHeight; ++row)
-		for (size_t col{ 0 }; col < mLevelByteWidth; ++col)
+	for (int row{ 0 }; row < mLevelHeight; ++row)
+		for (int col{ 0 }; col < mLevelByteWidth; ++col)
 		{
 			unsigned char byte{ (unsigned char)mpData[level * mLevelByteWidth * mLevelHeight + row * mLevelByteWidth + col] };
 			for (int bit{ 0 }; bit < 8; ++bit)
 				pLayout[row * mLevelWidth + col * 8 + bit] = (byte >> (7 - bit)) & 1;
 		}
-	for (size_t row{ 0 }; row < mLevelHeight; ++row)
+	for (int row{ 0 }; row < mLevelHeight; ++row)
 	{
 		pLayout[row * mLevelWidth + 30] = 1;
 		pLayout[row * mLevelWidth + 31] = 1;
